+++ app/src/main/java/com/cornellappdev/scoop/models/Search.kt
@@ -1,0 +1,7 @@
+package com.cornellappdev.scoop.models
+
+data class Search(
+    var departureLocation: String? = null,
+    var arrivalLocation: String? = null,
+    var departureDate: String? = null
+)
